+++ Assets/_Scripts/Bots/BotController.cs
@@ -8,32 +8,31 @@ namespace _Scripts.Bots
     {
         private enum BotMode
         {
-            Default,        // none of the below, move to random position around spawn location and enter idle animation on random stops
-            SearchEnemy,    // enemy is in detection radius, walk slowly towards random position near enemy
-            ChaseEnemy,     // enemy in detection radius and in field of view, run towards enemy position
-            AttackEnemy,    // enemy in attack range, stop movement and hit enemy
-            Die             // bot life == 0, play death animation and destroy gameobject
+            Default, // none of the below, move to random position around spawn location and enter idle animation on random stops
+            SearchEnemy, // enemy is in detection radius, walk slowly towards random position near enemy
+            ChaseEnemy, // enemy in detection radius and in field of view, run towards enemy position
+            AttackEnemy, // enemy in attack range, stop movement and hit enemy
+            Die // bot life == 0, play death animation and destroy gameobject
         }
 
         // bot behaviour
         private BotMode _mode = BotMode.Default;
-        
+
         private BotStats _stats;
-        
+
         [SerializeField] private float searchRadius = 10;
         [SerializeField] private float chaseRadius = 5;
         [SerializeField] public float attackRadius = 1;
         [SerializeField] private LayerMask mask;
-        
+
         // bot movement
         [SerializeField] private float stopMargin = 0.1f;
-        private float _stoppingDistance;
         private NavMeshAgent _agent;
         private Vector3 _destination;
 
         // bot animation
         private Animator _animator;
-        
+
         private int _isIdlingHash;
         private int _isWalkingHash;
         private int _isRunningHash;
@@ -42,17 +41,16 @@ namespace _Scripts.Bots
 
         // misc
         private Transform _target;
-        
+
         // Start is called before the first frame update
         void Start()
         {
             _animator = GetComponent<Animator>();
             _agent = GetComponent<NavMeshAgent>();
             _agent.updateRotation = true;
-            
+
             _agent.stoppingDistance = attackRadius - stopMargin;
-            _stoppingDistance = _agent.stoppingDistance;
-            
+
             _target = PlayerManager.instance.player.transform;
 
             _isIdlingHash = Animator.StringToHash("idle");
@@ -91,7 +89,6 @@ namespace _Scripts.Bots
                     Die();
                     break;
             }
-
         }
 
         /**
@@ -102,17 +99,16 @@ namespace _Scripts.Bots
         {
             var pos = transform.position;
             // _animator.SetBool(_isAttackingHash, false);
-            
+
             _mode = BotMode.Default;
-            // Debug.Log(DestinationReached());
-            
+
             // Check if Bot is dead
             if (_stats.CurrentHealth <= 0)
             {
                 _mode = BotMode.Die;
                 return;
             }
-            
+
             // Check if enemy in attack range
             if (Physics.CheckSphere(pos, attackRadius, mask))
             {
@@ -133,7 +129,7 @@ namespace _Scripts.Bots
 
                         if (angleToPlayer is >= -60 and <= 60) // 120° FOV 
                         {
-                            // Debug.Log("Player in sight!");
+                            Debug.Log("Player in sight!");
                             _mode = BotMode.ChaseEnemy;
                             return;
                         }
@@ -144,6 +140,7 @@ namespace _Scripts.Bots
             // Check if enemy in search range
             if (Physics.CheckSphere(pos, searchRadius, mask) || !DestinationReached())
             {
+                Debug.Log("searchRadius");
                 _mode = BotMode.SearchEnemy;
             }
         }
@@ -163,9 +160,10 @@ namespace _Scripts.Bots
          */
         private bool DestinationReached()
         {
-            var distance = Vector3.Distance(transform.position, _destination);
+            // var distance = Vector3.Distance(transform.position, _destination);
             // Debug.Log("Distance is: " + distance + " , _stoppingDistance is: " + _stoppingDistance);
-            return distance <= _stoppingDistance;
+            // return distance <= _stoppingDistance;
+            return _destination == transform.position;
         }
 
         /**
@@ -175,61 +173,55 @@ namespace _Scripts.Bots
          */
         private void Idling()
         {
-            // Debug.Log("Bot is in default mode");
+            Debug.Log("Bot is in default mode");
             _animator.SetTrigger(_isIdlingHash);
             // _animator.SetBool(isAttackingHash, false);
         }
 
         private void Search()
         {
+            _animator.SetTrigger(_isWalkingHash);
+            
             var pos = transform.position;
-            if (_destination == transform.position)
+    
+            Debug.Log("Bot is in search mode");
+            
+            if (Physics.CheckSphere(pos, searchRadius, mask))
             {
-                
+                Move(_target.position);
+            }
+            if (_destination != transform.position /* && Physics.CheckSphere(pos, searchRadius, mask)*/)
+            {
+                Move(_target.position);
+            }
+            else if (!DestinationReached() && !Physics.CheckSphere(pos, searchRadius, mask))
+            {
+                _animator.SetTrigger(_isWalkingHash);
+                Move(_target.position);
             }
             else
             {
-                // Debug.Log("Bot is in search mode");
-               _animator.SetTrigger(_isWalkingHash);
-               
-                           if (Physics.CheckSphere(pos, searchRadius, mask))
-                           {
-                               Move(_target.position);
-                           } else if (!DestinationReached() && Physics.CheckSphere(pos, searchRadius, mask))
-                           {
-                               Move(_target.position);
-                           } else if (DestinationReached() && Physics.CheckSphere(pos, searchRadius, mask))
-                           {
-                               Move(_target.position);
-                           }
-                           else
-                           {
-                               _destination = transform.position;
-                           } 
+                _destination = transform.position;
             }
-            
-
-            // if the old destination has been reached move to a new destination because player is in range
-            // if () Move(_target.position);            // only set new destination if the old one has been reached
         }
 
         private void Chase()
         {
-            // Debug.Log("Bot is in chase mode");
+            Debug.Log("Bot is in chase mode");
             _animator.SetTrigger(_isRunningHash);
             Move(_target.position);
         }
 
         private void Attack()
         {
-            // Debug.Log("Bot is in attack mode");
+            Debug.Log("Bot is in attack mode");
             _animator.SetBool(_isAttackingHash, true);
             FaceTarget();
         }
-        
+
         private void Die()
         {
-            // Debug.Log("Bot is in die mode");
+            Debug.Log("Bot is in die mode");
             if (!_animator.GetCurrentAnimatorStateInfo(0).IsName("Dying"))
             {
                 _animator.SetTrigger(_isDyingHash);